Implement opcodes until FX29

+++ includes/chip.hpp
@@ -17,9 +17,9 @@ class Chip {
     // 4KB memory, first 512 bytes are reserved
     // ROM is loaded after that
     std::vector<Byte> m_memory;
-    int m_programCounter;
+    unsigned short m_programCounter;
     // stores a 12 bit address pointer
-    int m_indexRegister;
+    unsigned short m_indexRegister;
     // V0 - VF
     std::vector<Byte> m_registers;
     // possibly 12 - 16 levels of pushing
@@ -30,8 +30,8 @@ class Chip {
     // 32 rows, 64 columns
     std::vector<bool> m_frameBuffer;
     // count down at 60hz unless they are 0
-    int m_delayTimer;
-    int m_soundTimer;
+    Byte m_delayTimer;
+    Byte m_soundTimer;
 
     // not part of chip8, useful for performance reasons
     // only draw when this flag is set
